pyproject.toml
[tool.poetry]
name = "fraud-detection"
version = "0.1.0"
description = ""
authors = ["patricijabautrenaite <p.bautrenaite@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
scikit-learn = "^1.4.2"
joblib = "^1.4.0"
pandas = "^2.2.2"
pyyaml = "^6.0.1"
mlflow = "^2.12.1"
imblearn = "^0.0"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
